Final Dask Code

+++ Dask/Generators/gen_indexed_csv.py
@@ -3,9 +3,12 @@ import sys
 import os
 import random
 
-def generate_data(filename,rows,cols, col_offset):
+def generate_data(filename,size,cols, col_offset):
+    if (os.path.exists(filename+".csv")):
+        return
     inputfile = open(filename+ ".csv", "w+")
-    for i in range(rows):
+    i = 0
+    while (os.stat(filename+".csv").st_size < size*1024*1024):
         for j in range(col_offset,cols+ col_offset):
             if i == 0:
                 if (j == cols + col_offset -1):
@@ -21,4 +24,9 @@ def generate_data(filename,rows,cols, col_offset):
                 else:
                     inputfile.write(lorem.sentence() + ",")
         inputfile.write("\n")
+        inputfile.flush()
+        i+=1
     inputfile.close()
+
+#generate_data("file1.csv",100,10,0)
+#generate_data("file2.csv",100,20,0)
